switch to simpler analyser plugins
